requirements.txt
supabase
python-telegram-bot
google-generativeai
python-dotenv
asyncio
requests
